requirements.txt
# Smart Feedback Analysis Platform - Essential Requirements

# Core Data Processing
pandas>=2.0.0
numpy>=1.20.0

# Natural Language Processing
nltk>=3.8.0
textblob>=0.17.0
scikit-learn>=1.3.0

# Database Connectivity (SQLite for simplicity)
sqlalchemy>=2.0.0

# Configuration and Utilities
pyyaml>=6.0.0
python-dotenv>=1.0.0
schedule>=1.2.0

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# HTTP Requests
requests>=2.31.0

# Testing
pytest>=7.4.0

# Logging
structlog>=23.0.0

# Date/Time Utilities
python-dateutil>=2.8.0
pytz>=2023.0

# Excel Support
openpyxl>=3.1.0

# Performance Monitoring
psutil>=5.9.0